# mnemosyne-web-interface/mnemosyne-api/docs/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/company/company-users:
    get:
      operationId: CompanyController_getCompanyUsers
      summary: Endpoint is responsible for getting the list of all users in company
        with pagination.
      parameters:
        - name: page
          required: true
          in: query
          description: Query for page in order to get list of company users.
          schema:
            type: string
        - name: pageSize
          required: true
          in: query
          description: Query for limit in order to get list of company users.
          schema:
            type: string
        - name: query
          required: false
          in: query
          description: Member query
          schema:
            type: string
      responses:
        "200":
          description: As the response user gets the list of all users along with count
            for pagination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompanyUsersDto"
        "400":
          description: Parse exception is thrown in case if page or limit params cannot be
            parsed as number.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseException"
      tags: *a4
      security:
        - x-access-token: []
        - basicAuth: []
...
  /api/company-users/invite-user:
    post:
      operationId: CompanyUsersController_inviteUser
      summary: Endpoint is responsible for user invitation to the company.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserToCompanyDto"
      responses:
        "201":
          description: As a response user gets message that user has been invited.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInvitedDto"
      tags: &a6
        - Company Users
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company-users/company-member-info:
    get:
      operationId: CompanyUsersController_getCompanyMemberInfo
      summary: Endpoint is responsible for obtaining the information about company
        member by ID.
      parameters:
        - name: memberId
          required: true
          in: query
          description: User (company member) ID.
          schema:
            type: string
      responses:
        "201":
          description: As a response user gets information about company user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMemberInfoDto"
        "404":
          description: In case if user ID has been modified and user not found, endpoint
            returns the error of user not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMemberNotFoundException"
      tags: *a6
      security:
        - x-access-token: []
        - basicAuth: []
    patch:
      operationId: CompanyUsersController_updateCompanyMemberInfo
      summary: Endpoint is responsible for updating the information of the company
        member.
      parameters:
        - name: memberId
          required: true
          in: query
          description: User (company member) ID.
          schema:
            type: string
      requestBody:
        required: true
        description: Body contains user personal information.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfoDto"
      responses:
        "201":
          description: As a response user gets message that information has been updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMemberInfoDto"
        "404":
          description: In case if user ID has been modified and user not found, endpoint
            returns the error of user not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMemberNotFoundException"
      tags: *a6
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company-users/delete-company-member:
    delete:
      operationId: CompanyUsersController_deleteCompanyMember
      summary: Endpoint is responsible for deleting user from the company.
      parameters:
        - name: memberId
          required: true
          in: query
          description: User (company member) ID.
          schema:
            type: string
      requestBody:
        required: true
        description: In the body user has to provide MFA code(s).
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCompanyMemberDto"
      responses:
        "201":
          description: In order to perform this action the MFA is required. Therefore,
            except the response that the user has been deleted, the user might
            get the response with the MFA requirement.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CompanyMemberDeletedDto"
                  - $ref: "#/components/schemas/FullMfaRequiredDto"
                  - $ref: "#/components/schemas/TokenTwoFaRequiredDto"
                  - $ref: "#/components/schemas/PhoneMfaRequiredDto"
        "404":
          description: In case if memberId was modified, or memberId of the different
            company was provided, endpoint returns not found error message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMemberNotFoundException"
      tags: *a6
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company-users/search-company-members:
    get:
      operationId: CompanyUsersController_searchCompanyMembers
      summary: Endpoint is responsible for searching company members.
      parameters:
        - name: query
          required: true
          in: query
          description: Member query
          schema:
            type: string
        - name: page
          required: true
          in: query
          description: Query for page in order to get list of company users.
          schema:
            type: string
        - name: pageSize
          required: true
          in: query
          description: Query for limit in order to get list of company users.
          schema:
            type: string
      responses:
        "201":
          description: As a response user gets the array of users with id and email.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchCompanyMembersDto"
      tags: *a6
      security:
        - x-access-token: []
        - basicAuth: []
...
info:
  title: Mnemosyne
  description: Documentation of Mnemosyne API.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-access-token:
      scheme: bearer
      bearerFormat: JWT
      description: "Please enter the JWT token in the following format: Bearer <JWT>"
      name: x-access-token
      type: apiKey
      in: header
    basicAuth:
      type: http
      scheme: basic
      description: Basic API authentication username and password
      in: header
  schemas:
...
    UpdateUserInfoDto:
      type: object
      properties:
        firstName:
          type: string
          description: First name
          example: Joe
        lastName:
          type: string
          description: Last name
          example: Doe
        namePronunciation:
          type: string
          description: This could be a phonetic pronunciation, or an example of something
            your name sounds like
          example: Zoe ("zo-ee")
        homeAddress:
          type: string
          description: User home address
          example: One Apple Park Way, United States
        homePhone:
          type: string
          description: User home phone
          example: (800) 1231-1234
...
    FullMfaRequiredDto:
      type: object
      properties:
        message:
          type: string
          description: Full MFA (token + phone) required response message
          example: full-mfa-required
      required:
        - message
    TokenTwoFaRequiredDto:
      type: object
      properties:
        message:
          type: string
          description: Token MFA required response message
          example: token-two-fa-required
      required:
        - message
    PhoneMfaRequiredDto:
      type: object
      properties:
        message:
          type: string
          description: Phone MFA required response message
          example: phone-required
      required:
        - message
...
    ParseException:
      type: object
      properties:
        message:
          type: string
          description: Parse exception
          example: parse-error
      required:
        - message
...
    GetCompanyUsersDto:
      type: object
      properties:
        companyUsers:
          description: List of company users (id and email)
          example:
            - id: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
              email: joe@doe.com
              role:
                id: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf8c
                name: PRIMARY_ADMIN
              registrationHash:
                confirmed: "false"
                createdAt: 2023-11-17T11:42:26.534Z
          type: array
          items:
            type: array
        count:
          type: number
          description: Count for the pagination
          example: "5"
      required:
        - companyUsers
        - count
...
    InviteUserToCompanyDto:
      type: object
      properties:
        invitedUsers:
          description: The list of invited users
          example:
            - email: joe@doe.com
              roleId: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf8c
              roleName: My Test Role
          type: array
          items:
            type: array
            minLength: 1
        language:
          type: string
          enum:
            - ru
            - pl
            - en
          description: One of 3 available languages
          example: pl
      required:
        - invitedUsers
    UserInvitedDto:
      type: object
      properties:
        message:
          type: string
          description: User invited response message
          example: user-invited
      required:
        - message
    CompanyMemberRole:
      type: object
      properties: {}
    CompanyMemberInfoDto:
      type: object
      properties:
        memberId:
          type: string
          description: User ID
          example: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
        email:
          type: string
          description: Email address
          example: joe@doe.com
        firstName:
          type: string
          description: First name
          example: Joe
        lastName:
          type: string
          description: Last name
          example: Doe
        namePronunciation:
          type: string
          description: This could be a phonetic pronunciation, or an example of something
            your name sounds like
          example: Zoe ("zo-ee")
        homeAddress:
          type: string
          description: User home address
          example: One Apple Park Way, United States
        homePhone:
          type: string
          description: User home phone
          example: (800) 1231-1234
        companyMemberRole:
          description: All information about company member role
          example:
            id: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf8c
            name: My Test Role
            description: My Test Role Description
            roleScopes:
              - USER_MANAGEMENT
          allOf:
            - $ref: "#/components/schemas/CompanyMemberRole"
      required:
        - memberId
        - email
        - firstName
        - lastName
        - namePronunciation
        - homeAddress
        - homePhone
        - companyMemberRole
    CompanyMemberNotFoundException:
      type: object
      properties:
        message:
          type: string
          description: Company member not found error message
          example: company-member-not-found
      required:
        - message
    DeleteCompanyMemberDto:
      type: object
      properties:
        phoneCode:
          type: string
          description: User phone code - MFA
          example: "123123"
        mfaCode:
          type: string
          description: User authentication application code - MFA
          example: "123123"
        language:
          type: string
          enum:
            - ru
            - pl
            - en
          description: One of 3 available languages
          example: pl
    CompanyMemberDeletedDto:
      type: object
      properties:
        message:
          type: string
          description: Company member deleted from the company response message
          example: company-member-deleted
      required:
        - message
    SearchCompanyMembersDto:
      type: object
      properties:
        companyMembers:
          description: Array with the id of the user and their emails
          example:
            - companyUserId: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
              email: joe@doe.com
          type: array
          items:
            type: array
      required:
        - companyMembers
...